Hotfix abs autofe operator

+++ src/upgini/autofe/unary.py
@@ -12,10 +12,12 @@ class Abs(PandasOperand, VectorizableMixin):
     group_index: int = 0
 
     def calculate_unary(self, data: pd.Series) -> pd.Series:
-        return data.abs()
+        return data.astype(np.float64).abs()
+        # return data.abs()
 
     def calculate_group(self, data: pd.DataFrame, **kwargs) -> pd.DataFrame:
-        return data.abs()
+        return data.astype(np.float64).abs()
+        # return data.abs()
 
 
 class Log(PandasOperand, VectorizableMixin):
